inference_service/requirements.txt
flask==2.3.3
tensorflow==2.14.0
pillow==10.0.1
numpy==1.24.3
prometheus-client==0.17.1
requests==2.31.0
gunicorn==21.2.0
python-multipart==0.0.6
flask-cors==4.0.0
psutil==5.9.5